\name{overlap.grpreg-package}
\alias{overlap.grpreg-package}
\docType{package}
\title{
%%\packageTitle{overlap.grpreg}
Regularization paths for regression models with grouped covariates
}
\description{
%%\packageDescription{overlap.grpreg}
Regularization paths for regression models with grouped covariates
}
\details{
%%The DESCRIPTION file:
%%\packageDESCRIPTION{overlap.grpreg}
%%\packageIndices{overlap.grpreg}

The following functions are documented in the help pages:
  \itemize{
    \item \code{\link{incidenceMatrix}}
    \item \code{\link{overlap.grpreg}}
    \item \code{\link{cv.overlap.grpreg}}
    \item \code{\link{plot.overlap.grpreg}}
 %%   \item \code{\link{plot.cv.overlap.grpreg}}
    \item \code{\link{predict.overlap.grpreg}}
    \item \code{\link{select.overlap.grpreg}}
    \item \code{\link{summary.cv.overlap.grpreg}}
  }
}

\author{Yaohui Zeng <yaohui-zeng@uiowa.edu>}
%%Maintainer: \packageMaintainer{Yaohui Zeng <yaohui-zeng@uiowa.edu>}

\references{
\itemize{
    \item Jacob, L., Obozinski, G., & Vert, J. P. (2009, June). Group lasso with overlap and graph lasso. \emph{In Proceedings of the 26th annual international conference on machine learning, ACM}: 433-440.
    \item Breheny, P. and Huang, J. (2009) Penalized methods for bi-level variable selection.  \emph{Statistics and its interface}, \strong{2}: 369-380. \url{myweb.uiowa.edu/pbreheny/publications/Breheny2009.pdf}
    \item Huang J., Breheny, P. and Ma, S. (2012). A selective review of group selection in high dimensional models. \emph{Statistical Science}, \strong{27}: 481-499. \url{myweb.uiowa.edu/pbreheny/publications/Huang2012.pdf}
    \item Breheny, P. and Huang, J. (to appear) Group descent algorithms for nonconvex penalized linear and logistic regression models with grouped predictors. \emph{Statistics and Computing}, to appear. \url{www.springerlink.com/openurl.asp?genre=article&id=doi:10.1007/s11222-013-9424-2}
    \item Breheny, P. The group exponential lasso for bi-level variable selection. In submission. \url{myweb.uiowa.edu/pbreheny/publications/grpexp.pdf}
  }
}

\keyword{package}
%%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}
\examples{
## see examples in overlap.grpreg
}
