\name{grpregOverlap-internal}
\title{Internal functions}
\alias{gamma2beta}
\description{Internal functions in the package.}
\usage{
gamma2beta(gamma, incidence.mat, grp.vec, family) 
}
\author{
Yaohui Zeng <yaohui-zeng@uiowa.edu>
}

\details{
This function is not intended for use by users. \code{gamma2beta} transforms the latent coefficient matrix (or vector) into non-latent form according to the grouping information.
}

\keyword{internal}
